prd_mcp_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-storage==2.18.2
google-cloud-discoveryengine==0.13.9
markdown==3.7
